% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapier-package.R
\docType{package}
\name{rapier-package}
\alias{rapier}
\alias{rapier-package}
\title{rapier: Extract the API of a Package}
\description{
Constructs a list of exported functions and methods for a package.
This is useful to verify if a package's API has changed between releases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/rapier}
  \item \url{http://krlmlr.github.io/rapier}
  \item Report bugs at \url{https://github.com/krlmlr/rapier/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
\keyword{internal}
